requirements.txt
psycopg2-binary==2.9.1
pika==1.2.0